# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FMD2r",
    "description": "",
    "license": {
      "name": "GPL-2.0-only",
      "identifier": "GPL-2.0-only"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/covers": {
      "get": {
        "tags": [
          "covers"
        ],
        "summary": "A manga cover, fetched with its module's cookies, user agent and referer and cached on disk.",
        "operationId": "getCover",
        "parameters": [
          {
            "name": "module",
            "in": "query",
            "description": "The website module the cover belongs to.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "url",
            "in": "query",
            "description": "The cover's URL on the site (http or https).",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "w",
            "in": "query",
            "description": "Scale down to this width in pixels (1 to 2000), keeping the aspect ratio.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The cover image",
            "content": {
              "image/*": {}
            }
          },
          "304": {
            "description": "The browser's copy (`If-None-Match`) is current"
          },
          "400": {
            "description": "Not an http(s) URL, or a private-network target",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Unknown module",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "502": {
            "description": "Upstream failed to deliver the cover",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/resolve": {
      "post": {
        "tags": [
          "series"
        ],
        "summary": "Find the module that handles a manga URL.",
        "operationId": "resolveUrl",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The module and the module-relative link of the series",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesRef"
                }
              }
            }
          },
          "404": {
            "description": "No module handles this URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/save-folder": {
      "post": {
        "tags": [
          "series"
        ],
        "summary": "The folder a download of the series would be saved in with the saved settings, the manga\nfolder included: what `POST /api/tasks` stores as the task's `save_to`\n(mangadownloader/forms/frmMain.pas:2685-2710).",
        "operationId": "saveFolder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveFolderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveFolder"
                }
              }
            }
          },
          "400": {
            "description": "Malformed body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/series": {
      "get": {
        "tags": [
          "series"
        ],
        "operationId": "getSeries",
        "parameters": [
          {
            "name": "module",
            "in": "query",
            "description": "Module ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "link",
            "in": "query",
            "description": "The series link relative to the module's `RootURL`.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The series' info and chapters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesInfo"
                }
              }
            }
          },
          "404": {
            "description": "No such module, or the module found no series there",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "502": {
            "description": "The website could not be reached",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tasks": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "The download queue, filtered, sorted and paged.",
        "operationId": "listTasks",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Only this status group.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/TaskGroup"
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "Text in the title, the module ID or a chapter name, ignoring case.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "description": "Unix ms: only tasks last downloaded (or, never downloaded, added) at or after this.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "Unix ms: only tasks last downloaded (or, never downloaded, added) at or before this.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "1-based page; default 1.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "Default 100, at most 1000.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Default: queue order.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/TaskSort"
            }
          },
          {
            "name": "desc",
            "in": "query",
            "description": "Sort descending.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "One page of the matching tasks",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskList"
                }
              }
            }
          },
          "400": {
            "description": "A malformed query, e.g. page 0",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Queue chapters of a series for download.",
        "operationId": "createTask",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewTask"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "The queued task",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSummary"
                }
              }
            }
          },
          "422": {
            "description": "No chapters, or no such module",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "tasks"
        ],
        "summary": "Remove the finished tasks, keeping their files (`RemoveAllFinishedTasks`,\nbaseunits/uDownloadsManager.pas:1987-2001).",
        "operationId": "removeFinishedTasks",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Must be `finished`.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/TaskGroup"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Removed"
          },
          "400": {
            "description": "`status` is not `finished`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChapterInfo": {
        "type": "object",
        "description": "One chapter of a series.",
        "required": [
          "name",
          "link",
          "downloaded"
        ],
        "properties": {
          "downloaded": {
            "type": "boolean",
            "description": "Whether the chapter was downloaded before, or marked when the series was added to the\nlibrary (mangadownloader/forms/frmMain.pas:2797-2846). The UI calls it \"seen\"."
          },
          "link": {
            "type": "string",
            "description": "Relative to the module's `RootURL`."
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "FieldProblem": {
        "type": "object",
        "description": "One rejected value of a 422.",
        "required": [
          "field",
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string"
          },
          "field": {
            "type": "string",
            "description": "The setting as a dotted path, like [`Problem::field`]."
          }
        }
      },
...
      "NewTask": {
        "type": "object",
        "description": "A download to queue: chapters of one series.",
        "required": [
          "module_id",
          "link",
          "title",
          "chapters"
        ],
        "properties": {
          "artists": {
            "type": "string",
            "description": "For `%ARTIST%` in the folder and chapter names."
          },
          "authors": {
            "type": "string",
            "description": "For `%AUTHOR%` in the folder and chapter names."
          },
          "chapters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NewTaskChapter"
            },
            "description": "In download order."
          },
          "link": {
            "type": "string",
            "description": "The series link relative to the module's `RootURL`."
          },
          "module_id": {
            "type": "string"
          },
          "save_to": {
            "type": "string",
            "description": "The folder the series folder is made in; the configured download folder when empty or\nmissing."
          },
          "title": {
            "type": "string"
          }
        }
      },
      "NewTaskChapter": {
        "type": "object",
        "description": "A chapter to download.",
        "required": [
          "name",
          "link"
        ],
        "properties": {
          "link": {
            "type": "string",
            "description": "Relative to the module's `RootURL`."
          },
          "name": {
            "type": "string"
          },
          "number": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "1-based position in the series' chapter list, for `%NUMBERING%`; the position in\n`chapters` when missing.",
            "minimum": 0
          }
        }
      },
...
      "Problem": {
        "type": "object",
        "description": "RFC 9457 problem details body.",
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string"
          },
          "field": {
            "type": [
              "string",
              "null"
            ],
            "description": "The setting a validation error (422) is about, as a dotted path such as\n`connections.timeout_secs` or `options.server`.\nKept for clients that read one field: the first of `fields`."
          },
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldProblem"
            },
            "description": "Every setting a validation error (422) is about, each with why it was rejected."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "The status code's reason phrase."
          },
          "type": {
            "type": "string",
            "description": "Always `about:blank`: the status code says it all."
          }
        }
      },
...
      "ResolveRequest": {
        "type": "object",
        "description": "A manga URL to resolve.",
        "required": [
          "url"
        ],
        "properties": {
          "url": {
            "type": "string"
          }
        }
      },
      "SaveFolder": {
        "type": "object",
        "description": "The folder a download is saved in.",
        "required": [
          "folder"
        ],
        "properties": {
          "folder": {
            "type": "string"
          }
        }
      },
      "SaveFolderRequest": {
        "type": "object",
        "description": "A download whose folder to show before it is queued.",
        "required": [
          "module_id",
          "title"
        ],
        "properties": {
          "artists": {
            "type": "string"
          },
          "authors": {
            "type": "string"
          },
          "module_id": {
            "type": "string"
          },
          "save_to": {
            "type": "string",
            "description": "The folder the user picked; the website's or the default destination when empty."
          },
          "title": {
            "type": "string"
          }
        }
      },
...
      "SeriesInfo": {
        "type": "object",
        "description": "A series as FMD2's info panel shows it.",
        "required": [
          "module_id",
          "link",
          "title",
          "alt_titles",
          "authors",
          "artists",
          "genres",
          "status",
          "summary",
          "chapters",
          "in_library"
        ],
        "properties": {
          "alt_titles": {
            "type": "string"
          },
          "artists": {
            "type": "string"
          },
          "authors": {
            "type": "string"
          },
          "chapters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChapterInfo"
            },
            "description": "In module order."
          },
          "cover_url": {
            "type": [
              "string",
              "null"
            ],
            "description": "The cover through `/api/covers`; `None` when the module reports none."
          },
          "genres": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The module's comma-separated genres, split."
          },
          "in_library": {
            "type": "boolean",
            "description": "Whether the series is a favorite."
          },
          "link": {
            "type": "string",
            "description": "The series link relative to the module's `RootURL`, as the module reports it."
          },
          "module_id": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/SeriesStatus"
          },
          "summary": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "SeriesRef": {
        "type": "object",
        "description": "A series: its module and its link relative to the module's `RootURL`.",
        "required": [
          "module_id",
          "link"
        ],
        "properties": {
          "link": {
            "type": "string"
          },
          "module_id": {
            "type": "string"
          }
        }
      },
      "SeriesStatus": {
        "type": "string",
        "description": "`MangaInfo.Status` (`MangaInfo_Status*`, baseunits/uBaseUnit.pas:230-233).",
        "enum": [
          "completed",
          "ongoing",
          "hiatus",
          "cancelled",
          "unknown"
        ]
      },
...
      "TaskCounts": {
        "type": "object",
        "description": "Tasks per status group, for the queue's group headers.",
        "required": [
          "downloading",
          "waiting",
          "stopped",
          "finished"
        ],
        "properties": {
          "downloading": {
            "type": "integer",
            "format": "int64",
            "description": "Preparing, downloading, converting or compressing.",
            "minimum": 0
          },
          "finished": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "stopped": {
            "type": "integer",
            "format": "int64",
            "description": "Stopped, failed or disabled.",
            "minimum": 0
          },
          "waiting": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "TaskGroup": {
        "type": "string",
        "description": "The status groups of the queue page.",
        "enum": [
          "downloading",
          "waiting",
          "stopped",
          "finished"
        ]
      },
      "TaskList": {
        "type": "object",
        "description": "One page of the queue.",
        "required": [
          "items",
          "total",
          "page",
          "per_page",
          "counts"
        ],
        "properties": {
          "counts": {
            "$ref": "#/components/schemas/TaskCounts",
            "description": "Tasks per group matching the text and date filters (not the status filter)."
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskSummary"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32",
            "description": "1-based.",
            "minimum": 0
          },
          "per_page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "description": "Tasks matching the filter, over all pages.",
            "minimum": 0
          }
        }
      },
...
      "TaskSort": {
        "type": "string",
        "description": "The columns FMD2 sorts its downloads list by (`CompareTaskContainer`,\nbaseunits/uDownloadsManager.pas:2046-2083), plus the queue order itself.",
        "enum": [
          "queue",
          "title",
          "status",
          "progress",
          "speed",
          "website",
          "save_to",
          "added"
        ]
      },
...
      "TaskSummary": {
        "type": "object",
        "description": "A task as the queue lists it.",
        "required": [
          "id",
          "module_id",
          "link",
          "title",
          "save_to",
          "status",
          "enabled",
          "running",
          "chapters",
          "chapter_count",
          "chapters_done",
          "current_chapter",
          "done",
          "total",
          "bytes_per_sec",
          "date_added"
        ],
        "properties": {
          "bytes_per_sec": {
            "type": "number",
            "format": "double",
            "description": "Download speed; 0 when not running."
          },
          "chapter_count": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "chapters": {
            "type": "string",
            "description": "The chapter the task is at, e.g. `Ch. 1101 (2/2)`."
          },
          "chapters_done": {
            "type": "integer",
            "format": "int32",
            "description": "Chapters downloaded.",
            "minimum": 0
          },
          "current_chapter": {
            "type": "integer",
            "format": "int32",
            "description": "Index of the chapter the task is at.",
            "minimum": 0
          },
          "date_added": {
            "type": "string",
            "description": "RFC 3339."
          },
          "date_last_downloaded": {
            "type": [
              "string",
              "null"
            ],
            "description": "RFC 3339 end of the last download run."
          },
          "done": {
            "type": "integer",
            "format": "int64",
            "description": "Pages of the current chapter done (FMD2's `DownCounter`).",
            "minimum": 0
          },
          "enabled": {
            "type": "boolean",
            "description": "Disabled tasks are skipped by start and start-all."
          },
          "error": {
            "type": [
              "string",
              "null"
            ],
            "description": "Why the task failed."
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "link": {
            "type": "string",
            "description": "The series link relative to the module's `RootURL`."
          },
          "module_id": {
            "type": "string"
          },
          "running": {
            "type": "boolean",
            "description": "Whether the task's thread runs (it may still show Waiting until its first chapter starts)."
          },
          "save_to": {
            "type": "string",
            "description": "The folder the chapters are saved in."
          },
          "status": {
            "$ref": "#/components/schemas/TaskState"
          },
          "title": {
            "type": "string"
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "description": "Pages of the current chapter.",
            "minimum": 0
          }
        }
      },
...
    }
  }
}